Updates
So I added a ton of stuff. I rearranged the logo it is correct placement. By removing it from the nav section. (quite silly that took me so long to figure it out). I also updated the css style sheets with transition codes for buttons. In addition, I also added and SVG file to the index.html with the animation styling to the CSS file.
I center the footer section to make it more align with the who page.